use more reliable state machine and cleanup
