+++ Cargo.toml
@@ -14,6 +14,6 @@ repository = "https://github.com/ostenning/rotary-encoder-embedded"
 embedded-hal = { version = "0.2.7", features = ["unproven"] }
 
 [features]
-full-step = []
-debounced = []
+default = ["standard"]
+standard = []
 angular-velocity = []
