+++ README.md
@@ -18,12 +18,11 @@ The `RotaryEncoder` can operate in a number of different modes, these modes prov
 The following modes are currently provided:
 
 | Feature flag  | Mode           | Desc.  |
-| ------------- |-------------| -----|
-| `full-step`        | `FullStepMode`              | Uses a full-step state machine for transitions |
-| `debounced`        | `DebouncedMode`             | Uses a half-step state machine with dynamic debouncing |
-| `angular-velocity` | `AngularVelocityMode`       | Uses a full-step state machine with additional angular-velocity calculations |
+| ------------- |----------------| -------|
+| `standard`         | `StandardMode`              | Uses a state machine for transitions |
+| `angular-velocity` | `AngularVelocityMode`       | Same as `standard` but with additional angular-velocity calculations |
 
-## `FullStepMode` example
+## `StandardMode` example
 
 ```rust
 fn main() -> ! {
@@ -34,141 +33,31 @@ fn main() -> ! {
     let mut rotary_encoder = RotaryEncoder::new(
         rotary_dt,
         rotary_clk,
-    ).into_fullstep_mode();
-
-    // Application loop
-    loop {
-        // Update the encoder, which will compute its direction
-        rotary_encoder.update();
-        match rotary_encoder.direction() {
-            Direction::Clockwise => {
-                // Increment some value
-            }
-            Direction::AntiClockwise => {
-                // Decrement some value
-            }
-            Direction::None => {
-                // Do nothing
-            }
-        }
-    }
-}
-```
-
-## `DebounceMode` example
-
-Trigger GPIO pin interrupts for both `DT` and `CLK` on both rising and falling edges
-
-```rust
-static ROTARY_ENCODER: Mutex<RefCell<Option<RotaryEncoder<_, _, FullStepMode>>>> = Mutex::new(RefCell::new(None));
-
-fn main() -> ! {
-    // Configure DT typically as pullup input & interrupt on rising/falling edges
-    let mut rotary_dt = rotary_dt.into_pull_up_input();
-    rotary_dt.make_interrupt_source(sys_cfg);
-    rotary_dt.trigger_on_edge(exti, Edge::RisingFalling);
-    rotary_dt.enable_interrupt(exti);
-    // Configure CLK typically as pullup input & interrupt on rising/falling edges
-    let mut rotary_clk = rotary_clk.into_pull_up_input();
-    rotary_clk.make_interrupt_source(sys_cfg);
-    rotary_clk.trigger_on_edge(exti, Edge::RisingFalling);
-    rotary_clk.enable_interrupt(exti);
-
-    // ... Configure a timer to interrupt at 60Hz
-
-    // Initialize Rotary Encoder and safely store in static global
-    interrupt::free(|cs| {
-        ROTARY_ENCODER.borrow(cs).replace(Some(
-            RotaryEncoder::new(
-                rotary_dt,
-                rotary_clk,
-            ).into_debounced_mode()
-        ));
-    });
-
-    
+    ).into_standard_mode();
+    // ...timer initialize at 900Hz to poll the rotary encoder
     loop {}
 }
 
-/// Called from Timer interrupt vector
-fn handle_timer_interrupt() {
-    interrupt::free(|cs| {
-        if let Some(ref mut rotary_encoder) = ROTARY_ENCODER.borrow(cs).borrow_mut().deref_mut() {
-          const SAMPLE_PERIOD: f32 = 1.0 / 60.0;
-          rotary_encoder.tick(SAMPLE_PERIOD);
+fn timer_interrupt_handler() {
+    // ... get rotary encoder 
+    let rotary_encoder = ...
+    // Update the encoder, which will compute its direction
+    rotary_encoder.update();
+    match rotary_encoder.direction() {
+        Direction::Clockwise => {
+            // Increment some value
         }
-    });
-}
-
-/// Called from the GPIO interrupt vector
-fn handle_rotary_interrupt() {
-    interrupt::free(|cs| {
-        if let Some(ref mut rotary_encoder) = ROTARY_ENCODER.borrow(cs).borrow_mut().deref_mut() {
-            // Borrow the pins to clear the pending interrupt bit (which varies depending on HAL)
-            let mut pins = rotary_encoder.pins_mut();
-            pins.0.clear_interrupt_pending_bit();
-            pins.1.clear_interrupt_pending_bit();
-            
-            // Update the encoder, which will compute its direction
-            // current_time should be a monotonously rising time in ms (akin to Arduino's `millis()`)
-            rotary_encoder.update(current_time);
-            
-            match rotary_encoder.direction() {
-                Direction::Clockwise => {
-                    // Increment some value
-                }
-                Direction::AntiClockwise => {
-                    // Decrement some value
-                }
-                Direction::None => {
-                    // Do nothing
-                }
-            }
+        Direction::AntiClockwise => {
+            // Decrement some value
         }
-    });
-}
-```
-
-## `AngularVelocityMode` example
-
-```rust
-fn main() -> ! {
-    // Configure DT and CLK pins, typically pullup input
-    let rotary_dt = gpio_pin_1.into_pull_up_input()
-    let rotary_clk = gpio_pin_2.into_pull_up_input();
-    // Initialize Rotary Encoder with Velocity functionality
-    let mut rotary_encoder = RotaryEncoder::new(
-        rotary_dt,
-        rotary_clk,
-    ).into_angular_velocity_mode();
-    // Optional settings
-    rotary_encoder.set_sensitivity(Sensitivity::Low);
-    rotary_encoder.set_velocity_action_ms(5);       // The window of time that the velocity may increase
-    rotary_encoder.set_velocity_inc_factor(0.2);    // How quickly the velocity increases over time
-    rotary_encoder.set_velocity_dec_factor(0.01);   // How quickly the velocity decreases over time
-    // Application loop
-    loop {
-        // Update the encoder which will compute its direction and velocity
-        // As velocity is a function of time, we need the current time
-        // current_time should be a monotonously rising time in ms (akin to Arduino's `millis()`)
-        rotary_encoder.update(current_time);
-        // Get the velocity
-        let velocity = rotary_encoder.velocity();
-        // Match the direction
-        match rotary_encoder.direction() {
-            Direction::Clockwise => {
-                // Increment some value by some factor multiplied by velocity
-            }
-            Direction::AntiClockwise => {
-                // Decrement some value by some factor multiplied by velocity
-            }
-            Direction::None => {
-                // Do nothing
-            }
+        Direction::None => {
+            // Do nothing
         }
-        // As velocity is a function of time, we need to reduce its value over time
-        // This method could also be called from a Timer
-        rotary_encoder.decay_velocity();
     }
 }
 ```
+
+# A note about GPIO or Timer interrupt usage
+
+I've experimented a lot with different combinations in order to make Rotary Encoders behave predictably because generally speaking they are fickle at best. From my experimentation I've learnt that using GPIO pin based interrupts generally isn't a good idea because they are more prone to noise and increase the risk of misfires and jumps.
+Timers on the other hand provide a low pass filtering quality because they don't pick up higher frequency switching that GPIO interrupts do. I have found that using a Timer between 850-1000Hz seems to work best.
